Move split files
