+++ index.js
@@ -1,18 +1,23 @@
-const { addSplitPointsAll } = require("./src/split");
-const { cornerPointMerger } = require("./src/cornerPointMerger");
-const { generateCornerPoints } = require("./src/split");
-const { generatePointSubset } = require("./src/split");
+const {
+  addSplitPointsAll,
+  generateCornerPoints,
+  generatePointSubset,
+  split,
+} = require("./src/splitTiles/split");
+const { cornerPointMerger } = require("./src/splitTiles/cornerPointMerger");
 const { mergeTiles } = require("./src/mergeTiles");
-const { split } = require("./src/split");
 const { sphericSplit, createSphericGrid } = require("./src/sphericSplit");
 
 module.exports = {
+  // Main process functions for splitting
   addSplitPoints: addSplitPointsAll,
   cornerPointMerger,
   generateCornerPoints,
   generatePointSubset,
-  mergeTiles,
   split,
+  // Main process functions for merging
+  mergeTiles,
+  // Final function for spheric splitting
   sphericSplit,
   createSphericGrid,
 };
