+++ src/process/consoleManager.js
