+++ src/splitTiles/addSplitPoints.js
@@ -1,5 +1,5 @@
-const { getSplitPoints, pushArray } = require("./utils");
-const { C, RUN_STATE } = require("./consoleManager");
+const { getSplitPoints, pushArray } = require("../utils");
+const { C, RUN_STATE } = require("../process/consoleManager");
 
 /**
  * @param {*} coordinates
